refactor: unify websocket implementation
